zipinfo_greasemonkey.js: More code golf
The "headerName" variable is reused (wrapped in "\r\n" and ": ").
So it does not really make sense to call it "headerName". "header"
is better, and the shortness also allows three lines to be collapsed
into one.
Trivia: Now the GreaseMonkey implementation is only three lines longer
than the XMLHttpRequest implementation (not that it matters).